requirements.txt

tensorflow
streamlit
pillow
